First commit for the gradient shader
This is work in progress towards having the gradient shader.

Added Gradient material, shader and scene example.

Changed float4 with half4 in the sky shader.

Added script to replace tabs with spaces.

+++ Assets/Shaders/GradientSkyShader.shader
@@ -102,7 +102,7 @@ Shader "Shaders101/GradientSkyShader"
                 return outData;
             }
             
-            float4 FragmentMain(InterpolatedData inData) : SV_Target
+            half4 FragmentMain(InterpolatedData inData) : SV_Target
             {
                 return inData.m_color;
             }
